switch to double precision

+++ src/main.cpp
@@ -4,16 +4,16 @@
 
 using namespace std;
 
-void mandelbrot(float c_real, float c_imaginary, int &iterations, float &real, float &imag)
+void mandelbrot(double c_real, double c_imaginary, int &iterations, double &real, double &imag)
 {
 
   const int num_iterations = 400;
-  const float max_radius = 10000;
+  const double max_radius = 10000;
 
 
-  complex<float> z = complex<float>(0,0);
-  complex<float> z_next = complex<float>(0,0);
-  const complex<float> c = complex<float>(c_real, c_imaginary);
+  complex<double> z = complex<double>(0,0);
+  complex<double> z_next = complex<double>(0,0);
+  const complex<double> c = complex<double>(c_real, c_imaginary);
 
   iterations = 0;
   while(iterations < num_iterations) {
@@ -32,10 +32,10 @@ void mandelbrot(float c_real, float c_imaginary, int &iterations, float &real, f
   imag = z.imag();
 
 }
-void Linspace(float* arr, float begin, float end, int N)
+void Linspace(double* arr, double begin, double end, int N)
 {
-  float dx = (end - begin) / N;
-  float value = begin;
+  double dx = (end - begin) / N;
+  double value = begin;
   for(int i=0; i < N; i++) {
     arr[i] = value;
     value+=dx;
@@ -76,17 +76,17 @@ struct Application
   sf::Texture* texture;
   sf::Sprite* sprite;
   int mouseDragging = 0;
-  float center_x;
+  double center_x;
   int center_col_delta;
-  float center_y;
+  double center_y;
   int center_row_delta;
   int mouse_down_col;
   int mouse_down_row;
-  float x_span;
-  float* x;
-  float* y;
-  float delta_linspace_x;
-  float delta_linspace_y;
+  double x_span;
+  double* x;
+  double* y;
+  double delta_linspace_x;
+  double delta_linspace_y;
 
   Application()
   {
@@ -96,8 +96,8 @@ struct Application
     delta_linspace_x = 0;
     delta_linspace_y = 0;
 
-    x = new float[W];
-    y = new float[H];
+    x = new double[W];
+    y = new double[H];
     x_span = 3;
     center_x = 0;
     center_y = 0;
@@ -142,8 +142,8 @@ struct Application
         if(event.type == sf::Event::MouseWheelScrolled) {
           cout << "mouse scrolling" << endl;
           cout << event.mouseWheelScroll.delta << endl;
-          float deltazoom = ((float)event.mouseWheelScroll.delta / 100);
-          float zoomlevel = 1 + deltazoom;
+          double deltazoom = ((double)event.mouseWheelScroll.delta / 100);
+          double zoomlevel = 1 + deltazoom;
           std::cout << "deltazoom" << " => " << deltazoom << std::endl;
           x_span *= zoomlevel;
         }
@@ -162,21 +162,21 @@ struct Application
         int delta_col = position.x - mouse_down_col;
         int delta_row = position.y - mouse_down_row;
 
-        delta_linspace_x = x_span * ((float)delta_row / (float)W);
-        delta_linspace_y = x_span * ((float)delta_col / (float)H);
+        delta_linspace_x = x_span * ((double)delta_row / (double)W);
+        delta_linspace_y = x_span * ((double)delta_col / (double)H);
 
       }
       // std::cout << "x_span" << " => " << x_span << std::endl;
       Linspace(x, -(x_span/2)-delta_linspace_x-center_x, (x_span/2)-delta_linspace_x-center_x, W);
       Linspace(y, -(x_span/2)-delta_linspace_y-center_y, (x_span/2)-delta_linspace_y-center_y, H);
 
-      // float max_mandelrot = 0;
+      // double max_mandelrot = 0;
       for(int i=0; i < W; i++)
         for(int j=0; j < H; j++) {
           // calculate divergence for mandelbrot set
           int iterations;
-          float real;
-          float imag;
+          double real;
+          double imag;
           mandelbrot(x[i], y[j],
               iterations, // OUT
               real, // OUT
